+++ src/format.task.ts
@@ -1,14 +1,22 @@
 import {printSpawnResult} from '@feltcoop/felt/util/process.js';
+import {z} from 'zod';
 
 import {TaskError, type Task} from './task/task.js';
 import {formatDirectory} from './format/formatDirectory.js';
 import {paths} from './paths.js';
-import type {FormatTaskArgs} from './formatTask.js';
-import {FormatTaskArgsSchema} from './formatTask.schema.js';
 
-export const task: Task<FormatTaskArgs> = {
+const Args = z
+	.object({
+		check: z
+			.boolean({description: 'exit with a nonzero code if any files are unformatted'})
+			.default(false),
+	})
+	.strict();
+type Args = z.infer<typeof Args>;
+
+export const task: Task<Args> = {
 	summary: 'format source files',
-	args: FormatTaskArgsSchema,
+	Args,
 	run: async ({args, log}) => {
 		const {check} = args;
 		// TODO forward prettier args