+++ src/task/logTask.ts
@@ -6,6 +6,7 @@ import {printValue} from '@feltcoop/felt/util/print.js';
 import type {ArgSchema, ArgsSchema} from '../utils/args.js';
 import {loadModules} from '../fs/modules.js';
 import {loadTaskModule, type TaskModuleMeta} from './taskModule.js';
+import {toVocabSchema} from '../utils/schema.js';
 
 export const logAvailableTasks = async (
 	log: Logger,
@@ -58,8 +59,10 @@ export const logTaskHelp = (log: Logger, meta: TaskModuleMeta): void => {
 	} = meta;
 	const printed: string[] = [];
 	printed.push(cyan(name), 'help', '\n' + task.summary || '(no summary available)');
-	if (task.args) {
-		const properties = toArgProperties(task.args);
+	if (task.Args) {
+		// TODO refactor to avoid using `toVocabSchema`, and then remove the `zodToJsonSchema` dep
+		const args = toVocabSchema(task.Args, 'Args') as ArgsSchema;
+		const properties = toArgProperties(args);
 		const longestTaskName = Math.max(
 			ARGS_PROPERTY_NAME.length,
 			toMaxLength(properties, (p) => p.name),