+++ src/typecheck.task.ts
@@ -1,14 +1,20 @@
-import {printSpawnResult, spawn} from '@feltcoop/felt/util/process.js';
+import {printSpawnResult, spawn, type SpawnResult} from '@feltcoop/felt/util/process.js';
+import {z} from 'zod';
 
 import {TaskError, type Task} from './task/task.js';
-import type {TypecheckTaskArgs} from './typecheckTask.js';
-import {TypecheckTaskArgsSchema} from './typecheckTask.schema.js';
 import {printCommandArgs, serializeArgs, toForwardedArgs} from './utils/args.js';
 import {sveltekitSync} from './utils/sveltekit.js';
 
-export const task: Task<TypecheckTaskArgs> = {
+const Args = z
+	.object({
+		tsconfig: z.string({description: 'path to tsconfig.json'}).default('tsconfig.json'),
+	})
+	.strict();
+type Args = z.infer<typeof Args>;
+
+export const task: Task<Args> = {
 	summary: 'typecheck the project without emitting any files',
-	args: TypecheckTaskArgsSchema,
+	Args,
 	run: async ({fs, args, log}): Promise<void> => {
 		const {tsconfig} = args;
 
@@ -19,10 +25,8 @@ export const task: Task<TypecheckTaskArgs> = {
 		const serializedTscArgs = ['tsc', ...serializeArgs(forwardedTscArgs)];
 		log.info(printCommandArgs(serializedTscArgs));
 		const tscTypecheckResult = await spawn('npx', serializedTscArgs);
-		if (!tscTypecheckResult.ok) {
-			throw new TaskError(`Failed to typecheck. ${printSpawnResult(tscTypecheckResult)}`);
-		}
 
+		let svelteCheckResult: SpawnResult | undefined;
 		if (await fs.exists('node_modules/.bin/svelte-check')) {
 			const forwardedSvelteCheckArgs = toForwardedArgs('svelte-check');
 			if (!forwardedSvelteCheckArgs.tsconfig) forwardedSvelteCheckArgs.tsconfig = tsconfig;
@@ -31,10 +35,19 @@ export const task: Task<TypecheckTaskArgs> = {
 				...serializeArgs(forwardedSvelteCheckArgs),
 			];
 			log.info(printCommandArgs(serializedSvelteCheckArgs));
-			const svelteCheckResult = await spawn('npx', serializedSvelteCheckArgs);
-			if (!svelteCheckResult.ok) {
-				throw new TaskError(`Failed to typecheck Svelte. ${printSpawnResult(svelteCheckResult)}`);
-			}
+			svelteCheckResult = await spawn('npx', serializedSvelteCheckArgs);
+		}
+
+		let errorMessage = '';
+		if (!tscTypecheckResult.ok) {
+			errorMessage = printSpawnResult(tscTypecheckResult);
+		}
+		if (svelteCheckResult && !svelteCheckResult.ok) {
+			if (errorMessage) errorMessage += ' ';
+			errorMessage += printSpawnResult(svelteCheckResult);
+		}
+		if (errorMessage) {
+			throw new TaskError(`Failed to typecheck. ${errorMessage}`);
 		}
 	},
 };